// src/gen/user.service.ts
export interface GetAllRelatedFriendRequest {
  userId: string;
}

export interface GetAllRelatedFriendResponse {
  friends: GetUserResponse[];
  metadata: ManageUserResponse | undefined;
}

export interface DeleteAddressRequest {
  id: string;
  userId: string;
}

export interface GetAddressRequest {
  id: string;
  userId: string;
}

export interface GetDefaultAddressRequest {
  userId: string;
}

export interface GetDefaultAddressResponse {
  address: Address | undefined;
  metadata: ManageUserResponse | undefined;
}

export interface GetAddressResponse {
  address: Address | undefined;
  metadata: ManageUserResponse | undefined;
}

export interface GetAddressesRequest {
  userId: string;
  limit: number;
  page: number;
}

export interface Address {
  id: string;
  isDefault: boolean;
  userId: string;
  title: string;
  address: string;
  type: string;
}

export interface GetAddressesResponse {
  addresses: Address[];
  total: number;
  metadata: ManageUserResponse | undefined;
}

export interface CreateAddressRequest {
  userId: string;
  title: string;
  address: string;
  type: string;
  isDefault: boolean;
}

export interface ManageAddressResponse {
  id: string;
  metadata: ManageUserResponse | undefined;
}

export interface UpdateAddressRequest {
  id: string;
  userId: string;
  title: string;
  address: string;
  type: string;
  isDefault: boolean;
}

export interface GetUserByUserNameRequest {
  username: string;
  password: string;
}

export interface IsTakenEmailRequest {
  email: string;
}

export interface IsTakenEmailResponse {
  isTaken: boolean;
}

export interface IsTakenPhoneNumberRequest {
  phoneNumber: string;
}

export interface IsTakenPhoneNumberResponse {
  isTaken: boolean;
}

export interface GetUserRequest {
  id: string;
}

export interface GetUserResponse {
  id: string;
  name: string;
  avatar: string;
  phoneNumber: string;
  country: string;
  address: string;
  state: string;
  city: string;
  zipCode: string;
  about: string;
  role: string;
  isPublic: boolean;
  email: string;
  gender: string;
  location: string;
}

export interface UpdateUserRequest {
  name: string;
  photoUrl: string;
  phoneNumber: string;
  country: string;
  address: string;
  state: string;
  city: string;
  zipCode: string;
  about: string;
  role: string;
  email: string;
  gender: string;
  id: string;
  location: string;
}

export interface ManageUserResponse {
  id: string;
  message: string;
  code: string;
  errMessage: string;
}

export interface CreateUserRequest {
  name: string;
  phoneNumber: string;
  email: string;
  gender: string;
  password: string;
  location: string;
}

export const USER_PROTO_SERVICE_PACKAGE_NAME = 'userProtoService';

export interface UserServiceClient {
  getUser(request: GetUserRequest): Observable<GetUserResponse>;

  updateUser(request: UpdateUserRequest): Observable<ManageUserResponse>;

  createUser(request: CreateUserRequest): Observable<ManageUserResponse>;

  isTakenEmail(request: IsTakenEmailRequest): Observable<IsTakenEmailResponse>;

  isTakenPhoneNumber(
    request: IsTakenPhoneNumberRequest,
  ): Observable<IsTakenPhoneNumberResponse>;

  getUserByUserName(
    request: GetUserByUserNameRequest,
  ): Observable<ManageUserResponse>;

  getAllRelatedFriend(
    request: GetAllRelatedFriendRequest,
  ): Observable<GetAllRelatedFriendResponse>;

  /** address */

  getAddresses(request: GetAddressesRequest): Observable<GetAddressesResponse>;

  createAddress(
    request: CreateAddressRequest,
  ): Observable<ManageAddressResponse>;

  updateAddress(
    request: UpdateAddressRequest,
  ): Observable<ManageAddressResponse>;

  deleteAddress(
    request: DeleteAddressRequest,
  ): Observable<ManageAddressResponse>;

  getAddress(request: GetAddressRequest): Observable<GetAddressResponse>;

  getDefaultAddress(
    request: GetDefaultAddressRequest,
  ): Observable<GetDefaultAddressResponse>;
}
// ...
export const USER_SERVICE_NAME = 'UserService';